backend/requirements.txt
langchain 
langchain_community
pinecone-client 
pinecone
openai 
torch
pypdf 
transformers 
faiss-cpu
python-dotenv
fastapi
PyPDF2